Create REXTRACT
A REXX script with a JCL driver, which was originally created to retrieve unique DB names from IMS mapping utility output. The user can extract any specific word or part of a word from any text input by specifying the beginning of the search expression and the length of the detail to be retrieved. The default delimiter is white space between words, but the user can also specify any additional separator.